requirements.txt
streamlit==1.14.0
pillow==9.2.0
